day_6_complete
completed day 6 which was on a website named reborgworld. Today I was able to learn a lot about functions on a simplified website. The programming was using python, although some of the functions were pre-made in order to speed up the learning process. I separated the different tasks by block quotes.