All files for the course
cloned the following files from: lenargasimov/100-days-of-python. In order to efficiently switch between the days of work. 

One of the things on my to-do list is to create a large repo, for people taking this course in order to also help them with efficiency/downtown of switching back and forth and loosing focus on programming and following the content.