day_9_cont
Had a busy week and this is the continuation of the 9th day of my programming journey. I am learning about dictionaries and how they operate through the use of keys and values. I can see how this can be useful when creating a table of data, or organizing data within a database.