day_11_capstone
Started working on capstone project today. I am focused on organizing myself before beginning this project. The tools that I have found are using a state machine in order to plan out the proper states of each task in my project. I am using an add-on named draw.io in order to map out the different events that will take place in the logic.