day_11_capstone_v2
Today I was able to finish most of the capstone project with a very blunt method. This was probably one of the most interesting projects that I have worked on since it was difficult to utilize the things that I am learning. I used the following standard methods including; dictionaries, lists, arrays, functions, flow control and logic diagrams to complete the project. It was probably the most I ever learned in a single sitting.