day_12_v1
In this program I am working on creating a guessing game. This is important to me since I used to play a game like this when I was younger. The game was called what are the odds which might be what I name this program lol. So the basic way to play the game is that there is a random number generated and you have a certain number of chances to guess the number.