day_12_complete
Finished the program about what are the odds. The premise of this program is to take a random number and try to guess that hidden number. At the beginning of the game the user is given an option to choose one of two paths. Either the 'easy' path or the 'hard' path. This determines how many lives the user will begin with.