day_13_complete
Completed day 13 of python debugging and it was clever and amazing. I think the best part of completing these lessons was actually figuring out how to think like a computer. Being able to think like a debugger was amazing. It was difficult to follow the code line by line but became easier the more I practiced.