day_14_complete
Today I finished the 14th day of the Udemy project. I was able to work with functions better and also program flow. This was the most difficult program for me to make work so far. But I was able to complete it after about 2 days. I also added some logic tables through drawio