day_15_complete
Completed day 15 of python course. I was able to fully integrate a coffee machine with many different methods. This was by far the most difficult program for me. I think the reason that I had so much trouble finishing this project was due to the fact that I did not have enough knowledge of functions and how they could be used. This is now not the case and I am actively working to improve on my skills.